API/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Pydantic para validación
pydantic==2.9.0
pydantic-settings==2.5.0

# Supabase
supabase==2.9.0
postgrest==0.17.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==3.2.2

# Utilidades
python-dotenv==1.0.0
httpx==0.27.0
email-validator==2.2.0

# Fechas y timezone
python-dateutil==2.9.0

# Testing (opcional para desarrollo)
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0

# Code quality (opcional)
black==24.8.0
flake8==7.1.0
mypy==1.11.0
